OAK-9843: secure facet query fails when the hits does not include data from some faceted field (#627)
* disable security pack in test (to avoid annoying logs)

* refactor: renamed SecureFacetCommonTest to FacetCommonTest. Removed tests duplicates in oak-search-elastic

* fix NPE in ElasticSecureFacetAsyncProvider
