+++ oak-search-elastic/src/main/java/org/apache/jackrabbit/oak/plugins/index/elastic/query/async/facets/ElasticSecureFacetAsyncProvider.java
@@ -17,6 +17,7 @@
 package org.apache.jackrabbit.oak.plugins.index.elastic.query.async.facets;
 
 import co.elastic.clients.elasticsearch.core.search.Hit;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.apache.jackrabbit.oak.plugins.index.elastic.query.ElasticRequestHandler;
 import org.apache.jackrabbit.oak.plugins.index.elastic.query.ElasticResponseHandler;
@@ -75,15 +76,15 @@ class ElasticSecureFacetAsyncProvider implements ElasticFacetProvider, ElasticRe
         final String path = elasticResponseHandler.getPath(searchHit);
         if (path != null && isAccessible.test(path)) {
             for (String field: facetFields) {
-                String value = searchHit.source().get(field).asText();
+                JsonNode value = searchHit.source().get(field);
                 if (value != null) {
                     facetsMap.compute(field, (column, facetValues) -> {
                         if (facetValues == null) {
                             Map<String, Integer> values = new HashMap<>();
-                            values.put(value, 1);
+                            values.put(value.asText(), 1);
                             return values;
                         } else {
-                            facetValues.merge(value, 1, Integer::sum);
+                            facetValues.merge(value.asText(), 1, Integer::sum);
                             return facetValues;
                         }
                     });
