--- oak-search-elastic/src/test/java/org/apache/jackrabbit/oak/plugins/index/elastic/ElasticSecureFacetCommonTest.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.jackrabbit.oak.plugins.index.elastic;
-
-import org.apache.jackrabbit.oak.Oak;
-import org.apache.jackrabbit.oak.jcr.Jcr;
-import org.apache.jackrabbit.oak.plugins.index.SecureFacetCommonTest;
-import org.apache.jackrabbit.oak.plugins.index.TestUtils;
-import org.junit.ClassRule;
-
-import javax.jcr.Repository;
-
-public class ElasticSecureFacetCommonTest extends SecureFacetCommonTest {
-
-    @ClassRule
-    public static final ElasticConnectionRule elasticRule =
-            new ElasticConnectionRule(ElasticTestUtils.ELASTIC_CONNECTION_STRING);
-
-    protected Repository createJcrRepository() {
-        indexOptions = new ElasticIndexOptions();
-        repositoryOptionsUtil = new ElasticTestRepositoryBuilder(elasticRule).build();
-        Oak oak = repositoryOptionsUtil.getOak();
-        Jcr jcr = new Jcr(oak);
-        return jcr.createRepository();
-    }
-
-    protected void assertEventually(Runnable r) {
-        TestUtils.assertEventually(r,
-                ((repositoryOptionsUtil.isAsync() ? repositoryOptionsUtil.defaultAsyncIndexingTimeInSeconds : 0) + ElasticIndexDefinition.BULK_FLUSH_INTERVAL_MS_DEFAULT) * 5);
-    }
-}
