+++ oak-search/src/test/java/org/apache/jackrabbit/oak/plugins/index/FacetCommonTest.java
@@ -24,7 +24,6 @@ import org.apache.jackrabbit.oak.plugins.index.search.util.IndexDefinitionBuilde
 import org.apache.jackrabbit.oak.query.AbstractJcrTest;
 import org.apache.jackrabbit.oak.query.facet.FacetResult;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -53,7 +52,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 
-public abstract class SecureFacetCommonTest extends AbstractJcrTest {
+public abstract class FacetCommonTest extends AbstractJcrTest {
     private static final Logger LOG = LoggerFactory.getLogger(AbstractJcrTest.class);
     private static final PerfLogger LOG_PERF = new PerfLogger(LOG);
     protected TestRepository repositoryOptionsUtil;
@@ -80,6 +79,7 @@ public abstract class SecureFacetCommonTest extends AbstractJcrTest {
         indexRule.property("cons").propertyIndex();
         indexRule.property("foo").propertyIndex().getBuilderTree().setProperty(FACET_PROP, true, Type.BOOLEAN);
         indexRule.property("bar").propertyIndex().getBuilderTree().setProperty(FACET_PROP, true, Type.BOOLEAN);
+        indexRule.property("baz").propertyIndex().getBuilderTree().setProperty(FACET_PROP, true, Type.BOOLEAN);
 
         indexOptions.setIndex(adminSession, indexName, builder);
         indexNode = indexOptions.getIndexNode(adminSession, indexName);
@@ -203,8 +203,6 @@ public abstract class SecureFacetCommonTest extends AbstractJcrTest {
         assertEventually(() -> assertEquals(actualAclLabelCount, getFacets()));
     }
 
-    //TODO Test is failing with lucene index.
-    @Ignore
     @Test
     public void statisticalFacets_withHitCountSameAsSampleSize() throws Exception {
         Node facetConfig = getOrCreateByPath(indexNode.getPath() + "/" + FACETS, "nt:unstructured", adminSession);
@@ -314,7 +312,7 @@ public abstract class SecureFacetCommonTest extends AbstractJcrTest {
         if (path != null) {
             pathCons = " AND ISDESCENDANTNODE('" + path + "')";
         }
-        String query = "SELECT [rep:facet(foo)], [rep:facet(bar)] FROM [nt:base] WHERE [cons] = 'val'" + pathCons;
+        String query = "SELECT [rep:facet(foo)], [rep:facet(bar)], [rep:facet(baz)] FROM [nt:base] WHERE [cons] = 'val'" + pathCons;
         Query q;
         QueryResult queryResult;
         try {
